+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.0.0] - 2025-10-19
 ### Added
 - Public Cloud Upload/Download files
 - Unity Cloud Provider implementation
@@ -49,7 +50,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/GameDataSystem/compare/0.5.1...main
+[Unreleased]: https://github.com/HyagoOliveira/GameDataSystem/compare/1.0.0...main
+[1.0.0]: https://github.com/HyagoOliveira/GameDataSystem/tree/1.0.0/
 [0.5.1]: https://github.com/HyagoOliveira/GameDataSystem/tree/0.5.1/
 [0.5.0]: https://github.com/HyagoOliveira/GameDataSystem/tree/0.5.0/
 [0.4.0]: https://github.com/HyagoOliveira/GameDataSystem/tree/0.4.0/